Implementacao parcial metodos que calculam estatisticas do diretorio que armazena arquivos

+++ go/virtualsf/model/stats.go
@@ -8,6 +8,6 @@ import (
 type VFStorageStats struct {
 	App       string
 	Date      time.Time
-	TotalSize float32
+	TotalSize int64
 	FileCount int
 }
